+++ .github/scripts/fetch_models.py
@@ -4,6 +4,23 @@ from urllib import request
 import zipfile
 import pickle
 
+def extract_zip(file, dest):
+    with zipfile.ZipFile(file, "r") as f:
+        f.extractall(dest)
+        file_list = f.namelist()
+    return file_list
+
+def get_model_name(file_path):
+    with zipfile.ZipFile(os.path.join(file_path), "r") as zip_file:
+        file_list = zip_file.namelist()
+        for fle in file_list:
+            if "_data.c" in fle:
+                base = os.path.basename(fle)
+                model_name = base.split("_data.c")[0]
+                return model_name
+        else:
+            raise Exception(f"Could not find name of root model for {file_path}")
+
 
 def main(args):
     url = "https://github.com/lharri73/PySimlink-ci-models/raw/master/generated/"
@@ -25,20 +42,28 @@ def main(args):
     models = []
     for file in files:
         print(f"Extracting {file}")
-        with zipfile.ZipFile(os.path.join(args.dir, "zips", file), "r") as zip_file:
-            zip_file.extractall(os.path.join(args.dir, "extract", file + "_e"))
-            file_list = zip_file.namelist()
+        file_list = extract_zip(os.path.join(args.dir, "zips", file), os.path.join(args.dir, "extract", file + "_e"))
+        
+        data_file = None
+        zip_file = None
         for fle in file_list:
-            if "_data.c" in fle:
-                base = os.path.basename(fle)
-                model_name = base.split("_data.c")[0]
-                models.append(model_name)
-                break
-        else:
-            raise Exception(f"Could not find name of root model for {file}")
+            if fle.split('.')[-1] == 'pkl':
+                data_file = fle
+            elif fle.split('.')[-1] == 'zip':
+                zip_file = fle
+        if zip_file is None or data_file is None:
+            raise Exception("invalid zip file format. Should contain data file and model zip")
+        model_name = get_model_name(os.path.join(args.dir, "extract", file + "_e", zip_file))
+        
+        models.append([
+            model_name,
+            os.path.join(args.dir, "extract", file + "_e", data_file),
+            os.path.join(args.dir, "extract", file + "_e", zip_file)
+        ])
 
     with open(os.path.join(args.dir, "manifest.pkl"), "wb") as f:
-        pickle.dump(list(zip(files, models)), f)
+        pickle.dump(models, f)
+
 
 
 if __name__ == "__main__":