+++ pysimlink/utils/model_utils.py
@@ -115,10 +115,10 @@ def mt_rebuild_check(model_paths: "anno.ModelPaths", force_rebuild: bool) -> boo
     """
     if not force_rebuild:
         return False
-
+    
     compile_info = os.path.join(model_paths.tmp_dir, "compile_info.pkl")
     if not os.path.exists(compile_info):
-        return False
+        return True
 
     with open(compile_info, "rb") as f:
         info = pickle.load(f)
@@ -132,3 +132,6 @@ def mt_rebuild_check(model_paths: "anno.ModelPaths", force_rebuild: bool) -> boo
         return tdiff > 1.0
     else:
         return True
+
+def sanitize_model_name(model_name):
+    return model_name.replace(" ", "").replace("-", "_").lower()